+++ android/app/src/main/java/com/on/turip/ui/common/base/BaseBottomSheetFragment.kt
@@ -1,0 +1,32 @@
+package com.on.turip.ui.common.base
+
+import android.os.Bundle
+import android.view.LayoutInflater
+import android.view.View
+import android.view.ViewGroup
+import androidx.viewbinding.ViewBinding
+import com.google.android.material.bottomsheet.BottomSheetDialogFragment
+
+abstract class BaseBottomSheetFragment<VB : ViewBinding> : BottomSheetDialogFragment() {
+    private var _binding: VB? = null
+    val binding get() = _binding!!
+
+    abstract fun inflateBinding(
+        inflater: LayoutInflater,
+        container: ViewGroup?,
+    ): VB
+
+    override fun onCreateView(
+        inflater: LayoutInflater,
+        container: ViewGroup?,
+        savedInstanceState: Bundle?,
+    ): View {
+        _binding = inflateBinding(inflater, container)
+        return binding.root
+    }
+
+    override fun onDestroyView() {
+        super.onDestroyView()
+        _binding = null
+    }
+}
